Fix tests file structure
